@@ -4,109 +4,54 @@
   "version": "0.0.1",
   "scripts": {
     "clean": "rm -rf ./static/g",
-    "sapper:dev": "sapper dev --ext '.svelte .md'",
-    "dev": "run-p --race sapper:dev tailwind:watch",
-    "copyFolder": "./scripts/copy_folder",
+    "copyFolder": "./scripts/copy_folder.cjs",
     "useCache": "npm run copyFolder -- --merge /opt/build/cache/hc_my_cache/static/g --to static/g",
     "fillCache": "npm run copyFolder -- --save static/g --to /opt/build/cache/hc_my_cache/static/g",
-    "build": "npm run clean && npm run tailwind:prod && sapper build --legacy --ext '.svelte .md'",
-    "export": "npm run tailwind:prod && NODE_ENV=production sapper export --legacy --ext '.svelte .md' --entry entries",
-    "netlify": "npm run useCache && npm run export && npm run fillCache",
-    "start": "node __sapper__/build",
-    "start:exported": "npx serve __sapper__/export",
-    "cy:run": "cypress run",
-    "cy:open": "cypress open",
-    "tailwind": "postcss static/tailwind.css -o static/index.css",
-    "tailwind:watch": "npm run tailwind -- -w",
-    "tailwind:prod": "NODE_ENV=production npm run tailwind",
-    "validate": "svelte-check --ignore src/node_modules/@sapper",
-    "test": "jest src",
-    "test:serve": "run-p --race dev cy:run"
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.[jt]s$": "babel-jest",
-      "^.+\\.svelte$": [
-        "svelte-jester",
-        {
-          "preprocess": true
-        }
-      ]
-    },
-    "moduleFileExtensions": [
-      "js",
-      "ts",
-      "svelte"
-    ],
-    "setupFilesAfterEnv": [
-      "@testing-library/jest-dom/extend-expect"
-    ]
+    "netlify": "npm run useCache && npm run build:noClean && npm run fillCache",
+    "dev": "svelte-kit dev",
+    "build": "npm run clean && svelte-kit build",
+    "build:noClean": "svelte-kit build",
+    "preview": "svelte-kit preview",
+    "check": "svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
+    "lint": "prettier --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
+    "format": "prettier --write --plugin-search-dir=. ."
   },
   "dependencies": {
     "compression": "^1.7.1",
     "js-yaml": "^3.13.1",
     "lodash": "^4.17.15",
-    "marked": "^0.7.0",
-    "polka": "next",
-    "sirv": "^1.0.0"
+    "marked": "^0.7.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/plugin-transform-runtime": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/preset-typescript": "^7.12.1",
-    "@babel/runtime": "^7.0.0",
-    "@fullhuman/postcss-purgecss": "^1.3.0",
-    "@rollup/plugin-babel": "^5.0.0",
-    "@rollup/plugin-commonjs": "^14.0.0",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^8.0.0",
-    "@rollup/plugin-replace": "^2.2.0",
-    "@rollup/plugin-typescript": "^6.0.0",
-    "@rollup/plugin-url": "^5.0.0",
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/svelte": "^3.0.0",
-    "@tsconfig/svelte": "^1.0.10",
-    "@types/compression": "^1.7.0",
-    "@types/jest": "^26.0.15",
-    "@types/lodash": "^4.14.136",
-    "@types/node": "^14.11.1",
-    "@types/polka": "^0.5.1",
-    "@types/testing-library__jest-dom": "^5.9.5",
-    "autoprefixer": "^9.7.4",
-    "babel-eslint": "^10.0.2",
-    "babel-jest": "^26.6.3",
-    "cssnano": "^4.1.10",
-    "eslint": "^6.1.0",
-    "eslint-config-prettier": "^6.0.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-svelte3": "^2.7.3",
-    "frontmatter": "0.0.3",
-    "jest": "^26.6.3",
+    "@rollup/plugin-replace": "^3.0.0",
+    "@sveltejs/adapter-static": "next",
+    "@sveltejs/kit": "next",
+    "@typescript-eslint/eslint-plugin": "^4.19.0",
+    "@typescript-eslint/parser": "^4.19.0",
+    "autoprefixer": "^10.2.6",
+    "cssnano": "^5.0.6",
+    "eslint": "^7.22.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-svelte3": "^3.2.0",
+    "frontmatter": "^0.0.3",
     "markdown-it": "^9.1.0",
     "mdsvex": "^0.8.8",
-    "npm-run-all": "^4.1.5",
-    "postcss-cli": "^6.1.3",
-    "prettier": "^2.1.2",
-    "prettier-plugin-svelte": "^1.4.1",
-    "rollup": "^2.3.4",
-    "rollup-plugin-svelte": "^6.0.0",
-    "rollup-plugin-terser": "^7.0.0",
-    "sapper": "^0.28.0",
-    "svelte": "^3.17.3",
-    "svelte-check": "^1.0.46",
-    "svelte-image": "happycollision/svelte-image#5952f2df",
-    "svelte-jester": "^1.1.5",
-    "svelte-preprocess": "^4.3.0",
-    "tailwindcss": "^1.9.6",
-    "tslib": "^2.0.1",
-    "typescript": "^4.0.3",
-    "xxhash": "^0.3.0"
+    "postcss": "^8.3.5",
+    "postcss-load-config": "^3.1.0",
+    "prettier": "~2.2.1",
+    "prettier-plugin-svelte": "^2.2.0",
+    "svelte": "^3.34.0",
+    "svelte-check": "^2.0.0",
+    "svelte-image": "happycollision/svelte-image#72ac687e53844289bcd23e73a22d44f7ac8c8ede",
+    "svelte-preprocess": "^4.7.3",
+    "tailwindcss": "^2.2.4",
+    "tslib": "^2.0.0",
+    "typescript": "^4.0.0"
   },
+  "type": "module",
   "volta": {
-    "node": "12.16.3",
-    "npm": "7.0.11"
+    "node": "16.4.0",
+    "npm": "7.18.1"
   }
 }
