@@ -1,10 +1,12 @@
 <footer class="bg-green-100 p-6 flex flex-col items-center">
   <div
-    class="flex flex-col items-center sm:flex-row sm:items-start mb-4 max-w-4xl ">
+    class="flex flex-col items-center sm:flex-row sm:items-start mb-4 max-w-4xl "
+  >
     <img
       class="flex-none mr-4 mb-4"
       src="/images/playhouse-pic.jpg"
-      alt="exterior of Post Playhouse" />
+      alt="exterior of Post Playhouse"
+    />
     <p>
       Post Playhouse, Inc. is Northwestern Nebraska's favorite live theatre
       company, producing several musicals running in a repertory schedule every
@@ -18,7 +20,8 @@
     <img
       class="flex-none w-24 sm:w-32 mr-4 mb-4"
       src="/images/nac.png"
-      alt="Nebraska Arts Council logo" />
+      alt="Nebraska Arts Council logo"
+    />
     <p class="text-xs">
       The Nebraska Arts Council, a state agency, has supported the programs of
       this organization through its matching grants program funded by the
@@ -26,14 +29,17 @@
       Cultural Endowment. Visit
       <a
         class="underline text-green-700"
-        href="http://www.nebraskaartscouncil.org">
+        href="http://www.nebraskaartscouncil.org"
+      >
         www.nebraskaartscouncil.org
       </a>
-      for information on how the Nebraska Arts Council can assist your
-      organization, or how you can support the Nebraska Cultural Endowment, at
+      for information on how the Nebraska Arts Council can assist your organization,
+      or how you can support the Nebraska Cultural Endowment, at
       <a
         class="underline text-green-700"
-        href="http://www.nebraskaculturalendowment.org">www.nebraskaculturalendowment.org</a>.
+        href="http://www.nebraskaculturalendowment.org"
+        >www.nebraskaculturalendowment.org</a
+      >.
     </p>
   </div>
 
@@ -41,15 +47,15 @@
     <img
       class="flex-none w-24 sm:w-32 mr-4 mb-4"
       src="/images/csc.png"
-      alt="Chadron State College logo" />
+      alt="Chadron State College logo"
+    />
     <p class="text-xs">
       <a class="underline text-green-700" href="https://csc.edu">
         Chadron State College
       </a>
-      is a proud sponsor of the Post Playhouse. CSC routinely donates lighting
-      equipment, sound equipment, and costumes to the theatre to help Post
-      Playhouse bring you the very best looking and sounding shows you'll see in
-      Western Nebraska.
+      is a proud sponsor of the Post Playhouse. CSC routinely donates lighting equipment,
+      sound equipment, and costumes to the theatre to help Post Playhouse bring you
+      the very best looking and sounding shows you'll see in Western Nebraska.
     </p>
   </div>
 
@@ -57,7 +63,8 @@
     <img
       class="flex-none w-24 sm:w-32 mr-4 mb-4"
       src="/images/ngpc.png"
-      alt="Nebraska Game and Parks logo" />
+      alt="Nebraska Game and Parks logo"
+    />
     <p class="text-xs">
       Post Playhouse Incorporated produces its summer season at the theatre on
       Fort Robinson State Park. The fort is run by the Nebraska Game and Parks
