@@ -219,7 +219,8 @@
     name="mc-embedded-subscribe-form"
     class="validate"
     target="_blank"
-    novalidate>
+    novalidate
+  >
     <div id="mc_embed_signup_scroll">
       <h2>Join our mailing list</h2>
       <div class="indicates-required">
@@ -236,7 +237,8 @@
           value=""
           name="EMAIL"
           class="required email"
-          id="mce-EMAIL" />
+          id="mce-EMAIL"
+        />
       </div>
       <div class="mc-field-group">
         <label for="mce-FNAME">
@@ -248,7 +250,8 @@
           value=""
           name="FNAME"
           class="required"
-          id="mce-FNAME" />
+          id="mce-FNAME"
+        />
       </div>
       <div class="mc-field-group">
         <label for="mce-LNAME">
@@ -260,17 +263,20 @@
           value=""
           name="LNAME"
           class="required"
-          id="mce-LNAME" />
+          id="mce-LNAME"
+        />
       </div>
       <div id="mce-responses" class="clear">
         <div
           class="response"
           id="mce-error-response"
-          style="display:none"></div>
+          style="display:none"
+        ></div>
         <div
           class="response"
           id="mce-success-response"
-          style="display:none"></div>
+          style="display:none"
+        ></div>
       </div>
       <!-- real people should not fill this in and expect good things - do not remove this or risk form bot signups-->
       <div style="position: absolute; left: -5000px;" aria-hidden="true">
@@ -278,7 +284,8 @@
           type="text"
           name="b_60731cccb7709051d790cb593_75ae084ea1"
           tabindex="-1"
-          value="" />
+          value=""
+        />
       </div>
       <input type="hidden" name="AQ" id="mce-AQ" value="website" />
       <div class="clear">
@@ -287,7 +294,8 @@
           value="Subscribe"
           name="subscribe"
           id="mc-embedded-subscribe"
-          class="button" />
+          class="button"
+        />
       </div>
     </div>
   </form>
