@@ -135,7 +135,7 @@
       {/each}
     </div>
 
-    {#each ['c', 'd', 'e', 'f', 'g', 'h', 'j', 'k', 'l', 'm'] as row}
+    {#each ["c", "d", "e", "f", "g", "h", "j", "k", "l", "m"] as row}
       <div data-row="{row}" class="row {row}">
         {#each [14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1] as i}
           <div class="seat">
