@@ -38,7 +38,8 @@
     <a
       class="link-green"
       href="http://www.bing.com/maps/?v=2&amp;cp=r3b57g6cb5p3&amp;lvl=18&amp;dir=0&amp;sty=b&amp;sp=Point.r3b4xz6cb5sc_Post%20Playhouse_Post%20Playhouse%20is%20located%20in%20Historic%20Fort%20Robinson%20State%20Park%2C%20and%20is%20Northwestern%20Nebraska's%20Premiere%20venue%20for%20live%20theatre._http%3A%2F%2Fwww.postplayhouse.com_http%3A%2F%2Fwww.postplayhouse.com%2Fwordpress%2Fwp-content%2Fthemes%2Fpostplayhouse2.1%2Fimages%2Fplayhouse%255fpic.jpg_"
-      target="_blank">
+      target="_blank"
+    >
       Bing Maps
     </a>
     --- to the theatre
@@ -47,7 +48,8 @@
     <a
       class="link-green"
       href="http://maps.google.com/maps/ms?f=q&amp;hl=en&amp;geocode=&amp;ie=UTF8&amp;msa=0&amp;ll=42.668963,-103.464013&amp;spn=0.003826,0.009656&amp;z=17&amp;iwstate1=dir&amp;msid=113724959975087314194.00045338a29dc15cb2d92"
-      target="_blank">
+      target="_blank"
+    >
       Google Maps
     </a>
     --- to the theatre
@@ -56,7 +58,8 @@
     <a
       class="link-green"
       href="http://www.mapquest.com/maps?state=NE&amp;cat=fort+robinson+state+park"
-      target="_blank">
+      target="_blank"
+    >
       MapQuest
     </a>
     --- to Fort Robinson
@@ -66,7 +69,8 @@
       class="link-green"
       href="http://maps.yahoo.com/#mvt=m&amp;lat=42.66546&amp;lon=-103.460484&amp;zoom=16&amp;tt=fort
       robinson state park&amp;tp=1"
-      target="_blank">
+      target="_blank"
+    >
       Yahoo Maps
     </a>
     --- to Fort Robinson
@@ -84,14 +88,16 @@
       scrolling="no"
       marginheight="0"
       marginwidth="0"
-      src="https://maps.google.com/maps/ms?f=q&amp;hl=en&amp;geocode=&amp;ie=UTF8&amp;msa=0&amp;msid=215815128770604665825.00045338a29dc15cb2d92&amp;ll=42.668963,-103.464013&amp;spn=0.003826,0.009656&amp;t=m&amp;output=embed"></iframe>
+      src="https://maps.google.com/maps/ms?f=q&amp;hl=en&amp;geocode=&amp;ie=UTF8&amp;msa=0&amp;msid=215815128770604665825.00045338a29dc15cb2d92&amp;ll=42.668963,-103.464013&amp;spn=0.003826,0.009656&amp;t=m&amp;output=embed"
+    ></iframe>
   </div>
 </div>
 <small>
   View
   <a
     href="https://maps.google.com/maps/ms?f=q&amp;hl=en&amp;geocode=&amp;ie=UTF8&amp;msa=0&amp;msid=215815128770604665825.00045338a29dc15cb2d92&amp;ll=42.668963,-103.464013&amp;spn=0.003826,0.009656&amp;t=m&amp;source=embed"
-    class="link-green">
+    class="link-green"
+  >
     the theatre
   </a>
   in a larger map
@@ -106,6 +112,7 @@
   minutes of show times, or at the Lodge nearby. The State officers usually
   patrol the park during show times, so please do not forget to purchase a pass.
   Parking passes can also be purchased at the
-  <a href="https://ngpc-permits.ne.gov/NGPC-PS/faces/public/welcome">Nebraska
-    Game and Parks Commission website</a>.
+  <a href="https://ngpc-permits.ne.gov/NGPC-PS/faces/public/welcome"
+    >Nebraska Game and Parks Commission website</a
+  >.
 </p>
